requirements.txt
pymongo==4.12.0
requests==2.32.3
python-dateutil==2.9.0
ics==0.7.2